renames housing component
